requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
Cython==3.0.8
et-xmlfile==1.1.0
flatbuffers==23.5.26
frozenlist==1.4.1
humanfriendly==10.0
idna==3.6
ImageHash==4.3.1
magic-filter==1.0.12
mpmath==1.3.0
multidict==6.0.5
mysql-connector-python==8.3.0
numpy==1.26.4
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
packaging==23.2
pandas==2.2.2
pillow==10.2.0
protobuf==4.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.5.0
schedule==1.2.1
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.27
sympy==1.12
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
yarl==1.9.4
